+++ Grimeron/src/org/arps/Grimeron/entity/Tile.java
@@ -30,7 +30,7 @@ public class Tile{
     private final int gameX;
     private final int gameY;
     
-    public double weight = (Player.Place.NONE.getWeight() + Player.Place.FIRST.getWeight()) / 2;
+    public float weight = (Player.Place.NONE.getWeight() + Player.Place.FIRST.getWeight()) / 2;
     
     private Color color;
